make window file
